Extract FakeNavigator
