+++ samples/star/src/test/kotlin/com/slack/circuit/star/petlist/PetListPresenterTest.kt
@@ -27,6 +27,7 @@ import com.slack.circuit.star.data.Links
 import com.slack.circuit.star.data.Photo
 import com.slack.circuit.star.petdetail.PetDetailScreen
 import com.slack.circuit.star.repo.PetRepository
+import com.slack.circuit.test.FakeNavigator
 import com.slack.circuit.test.test
 import kotlinx.coroutines.test.runTest
 import org.junit.Test
@@ -113,34 +114,4 @@ class PetListPresenterTest {
 class TestRepository(private val animals: List<Animal>) : PetRepository {
   override suspend fun getAnimals(): List<Animal> = animals
   override suspend fun getAnimal(id: Long): Animal? = animals.firstOrNull { it.id == id }
-}
-
-// TODO move this to test artifact
-class FakeNavigator : Navigator {
-  private val navigatedScreens = Turbine<Screen>()
-  private val pops = Turbine<Unit>()
-
-  override fun goTo(screen: Screen) {
-    navigatedScreens.add(screen)
-  }
-
-  override fun pop(): Screen? {
-    pops.add(Unit)
-    return null
-  }
-
-  // For non-coroutines users only
-  fun takeNextScreen() = navigatedScreens.takeItem()
-
-  suspend fun awaitNextScreen() = navigatedScreens.awaitItem()
-
-  suspend fun awaitPop(): Unit = pops.awaitItem()
-
-  fun assertIsEmpty() {
-    navigatedScreens.ensureAllEventsConsumed()
-  }
-
-  fun expectNoEvents() {
-    navigatedScreens.expectNoEvents()
-  }
 }
